Added Pause Menu Prefab
Added a pause menu prefab. After placing in scene, the menu can be opened with [esc] and buttons will either resume the game or take you to main menu. [esc] can also be pressed to close the menu and [m] will open it. Buttons do not currently work if the player is in the scene, but the keys will still function.