Created sound class and audioManager class, added win Logic
Audiomanager class is in the main scene, dont destroy on load, same with a background music game object. The audioManager class utilizes the sound class. Also added some win logic that respawns the player into the playerController script. I need to seperate the game logic from this script but for now it worked a lot easier to add this code to the playerController script.